+++ TravelerApp/TravelerAppCore/Controller/Menu.cs
@@ -1,25 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using TravelerAppCore.Controller;
 using TravelerAppCore.Models.Hotels;
 using TravelerAppCore.View;
 
 namespace TravelerAppConsole
 {
-    public class Option
-    {
-        private static int id = 0;
-        public int _id { get; set; }
-        public string _optionText { get; set; }
-        public Options _menuOptions { get; set; }
-        public Option(string optionText, Options options)
-        {
-            _optionText = optionText;
-            _menuOptions = options;
-            _id = id;
-            id++;
-        }
-    }
     public class Menu
     {
         public static List<Option> MenuList = new List<Option>();
@@ -52,7 +39,6 @@ namespace TravelerAppConsole
             }
         }
         static int i = 0;
-
         public void DisplayMenu()
         {
             Console.CursorVisible = false;
